Move EntitiesProcessor to common:util
